Searching People On Social Networking Sites using Face Recognition
Face recognition app for android – Two options provided – Either a Video
or an image. If a video is selected, the video is converted into frames
and then a face detection is done on the frames and only the frames with
faces are displayed in a grid view. Once the image is selected, the
image is sent to a server for face recognition and the response is sent
back to the server which would be the name of the person recognized in
the image. All this is done using a PHP script. Then the user is
searched on Facebook , Linkedin , Twitter and Google + using the name
which was returned after face recognition.